organized tests in different packages. Updated some UI tests.
