+++ Hestia/app/src/androidTest/java/com/rugged/application/hestia/UITests/LogoutTest.java
@@ -1,14 +1,11 @@
-package com.rugged.application.hestia;
+package com.rugged.application.hestia.UITests;
 
 import android.support.test.espresso.intent.rule.IntentsTestRule;
 import android.support.test.runner.AndroidJUnit4;
-
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-
 import hestia.UI.DeviceListActivity;
-
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.intent.Intents.intended;
